# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Agent-Builder API
  description: |
    Agent-Builder is a comprehensive platform for creating LLM-based agents through a five-phase workflow.

    ## Features
    - 🔐 SSO Authentication (Google, Azure AD, Okta)
    - 🤖 Multi-LLM Support (Claude, OpenAI, Gemini)
    - 📊 Real-time Progress Tracking via WebSocket
    - 🧠 Shared Learning System
    - 📦 Multiple Output Formats (MCP, CLI, Skill, Library)
    - 🔒 Secure API Key Management (AES-256-GCM)

    ## Authentication
    All API endpoints (except `/auth/*` and `/health`) require a JWT token obtained via OAuth flow.

    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - **Agent Creation**: 10 requests per hour
    - **API Key Operations**: 20 requests per hour
    - **General API**: 100 requests per minute
    - **Downloads**: 50 requests per hour

    ## Base URL
    - Production: `https://api.agent-builder.com`
    - Development: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: Agent-Builder Support
    email: support@agent-builder.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.agent-builder.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/google/callback:
    get:
      summary: Google OAuth callback
      description: Handles OAuth callback and returns JWT token
      tags:
        - Authentication
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/me:
    get:
      summary: Get current user
      description: Returns authenticated user information
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  session:
                    type: object
                    properties:
                      expiresAt:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/auth/logout:
    post:
      summary: Logout
      description: Invalidate current session
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/agents/create:
    post:
      summary: Create a new agent
      description: |
        Start agent creation workflow. This is an asynchronous operation that returns
        a session ID immediately. Monitor progress via WebSocket or polling the session endpoint.

        **Prerequisites:**
        - User must have a valid Anthropic API key configured

        **Workflow Phases:**
        1. Clarification (2-3 rounds)
        2. Design (with extended thinking)
        3. Implementation (parallel code + tests + docs)
        4. Packaging
        5. Learning

        **Average Duration:** 20-35 minutes
      tags:
        - Agents
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
            examples:
              mcpServer:
                summary: Create MCP Server
                value:
                  description: A GitHub repository analyzer that extracts metrics and insights
                  outputType: mcp
                  language: typescript
                  interactive: false
              pythonCLI:
                summary: Create Python CLI
                value:
                  description: A log file parser that extracts error patterns
                  outputType: cli
                  language: python
                  interactive: false
      responses:
        '202':
          description: Agent creation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: pending
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
        user:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        provider:
          type: string
          enum: [google, azure, okta]
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required:
        - description
      properties:
        description:
          type: string
          minLength: 10
          description: Natural language description of the agent to build
          example: A web scraper that extracts product prices from e-commerce sites
        outputType:
          type: string
          enum: [skill, mcp, cli, library]
          default: mcp
          description: Output format
        language:
          type: string
          enum: [typescript, python]
          default: typescript
          description: Programming language
        interactive:
          type: boolean
          default: false
          description: Enable interactive clarification mode
...
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid or expired token
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation Error
            message: Invalid request parameters
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Too Many Requests
            message: Rate limit exceeded. Please try again later.
      headers:
        Retry-After:
          schema:
            type: integer
            description: Seconds until rate limit resets